Simplify transcript controls layout.
The controls will now always be on one line. I also removed the button
to close the transcript screen for fear of making the control panel too
large on very small screens. There will still be a back button at the
top. Might decide to add it again in the future and turn the control
panel into a scrollable.